Replaces direct dependencies on std::string, iostreams, etc.
Looking ahead and trying to prevent changes of the string type from requiring
massive rework/search'n'replace throughout the system.  This change would
simplify the change to use wstring or a different custom type in the future.
It also causes appropriate standard in/out streams, regex type, and stringstream
to be selected based on the defined parka char_type.